+++ src/doc-cache.ts
@@ -24,11 +24,11 @@ import { Observable } from 'rxjs';
 declare type CacheItem<RxDocType, OrmMethods> = [
     /**
      * Store the different document states of time
-     * based on their revision height.
+     * based on their revision height (rev height = array index).
      * We store WeakRefs so that we can later clean up
      * document states that are no longer needed.
      */
-    Map<number, WeakRef<RxDocument<RxDocType, OrmMethods>>>,
+    WeakRef<RxDocument<RxDocType, OrmMethods>>[],
 
     /**
      * Store the latest known document state.
@@ -66,7 +66,13 @@ declare type FinalizationRegistryValue = {
  * @link https://caniuse.com/?search=weakref
  */
 export class DocumentCache<RxDocType, OrmMethods> {
-    public cacheItemByDocId = new Map<string, CacheItem<RxDocType, OrmMethods>>();
+    public readonly cacheItemByDocId = new Map<string, CacheItem<RxDocType, OrmMethods>>();
+
+    /**
+     * Process stuff lazy to not block the CPU
+     * on critical paths.
+     */
+    public readonly tasks = new Set<Function>();
 
     /**
      * Some JavaScript runtimes like QuickJS,
@@ -79,8 +85,8 @@ export class DocumentCache<RxDocType, OrmMethods> {
             const docId = docMeta.docId;
             const cacheItem = this.cacheItemByDocId.get(docId);
             if (cacheItem) {
-                cacheItem[0].delete(docMeta.revisionHeight);
-                if (cacheItem[0].size === 0) {
+                cacheItem[0][docMeta.revisionHeight] = undefined as any;
+                if (cacheItem[0].length === 0) {
                     /**
                      * No state of the document is cached anymore,
                      * so we can clean up.
@@ -93,20 +99,42 @@ export class DocumentCache<RxDocType, OrmMethods> {
 
     constructor(
         public readonly primaryPath: string,
-        public readonly changes$: Observable<RxChangeEvent<RxDocType>>,
+        public readonly changes$: Observable<RxChangeEvent<RxDocType>[]>,
         /**
          * A method that can create a RxDocument by the given document data.
          */
         public documentCreator: (docData: RxDocumentData<RxDocType>) => RxDocument<RxDocType, OrmMethods>
     ) {
-        changes$.subscribe(changeEvent => {
-            const docId = changeEvent.documentId;
-            const cacheItem = this.cacheItemByDocId.get(docId);
-            if (cacheItem) {
-                const documentData = getDocumentDataOfRxChangeEvent(changeEvent);
-                cacheItem[1] = documentData;
+        changes$.subscribe(events => {
+            this.tasks.add(() => {
+                const cacheItemByDocId = this.cacheItemByDocId;
+                for (let index = 0; index < events.length; index++) {
+                    const event = events[index];
+                    const cacheItem = cacheItemByDocId.get(event.documentId);
+                    if (cacheItem) {
+                        let documentData = event.documentData;
+                        if (!documentData) {
+                            documentData = event.previousDocumentData as any;
+                        }
+                        cacheItem[1] = documentData;
+                    }
+                }
+            });
+            if (this.tasks.size <= 1) {
+                requestIdlePromiseNoQueue().then(() => {
+                    this.processTasks();
+                });
             }
         });
+    }
+
+    public processTasks() {
+        if (this.tasks.size === 0) {
+            return;
+        }
+        const tasks = Array.from(this.tasks);
+        tasks.forEach(task => task());
+        this.tasks.clear();
     }
 
     /**
@@ -138,11 +166,13 @@ export class DocumentCache<RxDocType, OrmMethods> {
      * Throws if not exists
      */
     public getLatestDocumentData(docId: string): RxDocumentData<RxDocType> {
+        this.processTasks();
         const cacheItem = getFromMapOrThrow(this.cacheItemByDocId, docId);
         return cacheItem[1];
     }
 
     public getLatestDocumentDataIfExists(docId: string): RxDocumentData<RxDocType> | undefined {
+        this.processTasks();
         const cacheItem = this.cacheItemByDocId.get(docId);
         if (cacheItem) {
             return cacheItem[1];
@@ -172,11 +202,12 @@ function getCachedRxDocumentMonad<RxDocType, OrmMethods>(
             const docId: string = (docData as any)[primaryPath];
             const revisionHeight = getHeightOfRevision(docData._rev);
 
-            let byRev: Map<number, WeakRef<RxDocument<RxDocType, OrmMethods>>>;
+            // array, indexed by revision height number
+            let byRev: WeakRef<RxDocument<RxDocType, OrmMethods>>[];
             let cachedRxDocumentWeakRef: WeakRef<RxDocument<RxDocType, OrmMethods>> | undefined;
             let cacheItem = cacheItemByDocId.get(docId);
             if (!cacheItem) {
-                byRev = new Map();
+                byRev = [];
                 cacheItem = [
                     byRev,
                     docData
@@ -184,13 +215,13 @@ function getCachedRxDocumentMonad<RxDocType, OrmMethods>(
                 cacheItemByDocId.set(docId, cacheItem);
             } else {
                 byRev = cacheItem[0];
-                cachedRxDocumentWeakRef = byRev.get(revisionHeight);
+                cachedRxDocumentWeakRef = byRev[revisionHeight];
             }
             let cachedRxDocument = cachedRxDocumentWeakRef ? cachedRxDocumentWeakRef.deref() : undefined;
             if (!cachedRxDocument) {
                 docData = deepFreezeWhenDevMode(docData) as any;
                 cachedRxDocument = documentCreator(docData) as RxDocument<RxDocType, OrmMethods>;
-                byRev.set(revisionHeight, createWeakRefWithFallback(cachedRxDocument));
+                byRev[revisionHeight] = createWeakRefWithFallback(cachedRxDocument);
                 if (registry) {
                     registryTasks.push(cachedRxDocument);
                 }
@@ -203,7 +234,7 @@ function getCachedRxDocumentMonad<RxDocType, OrmMethods>(
              * really bad performance. So we add the cached documents
              * lazily.
              */
-            requestIdlePromiseNoQueue().then(() => {
+            docCache.tasks.add(() => {
                 for (let index = 0; index < registryTasks.length; index++) {
                     const doc = registryTasks[index];
                     registry.register(doc, {
@@ -212,6 +243,11 @@ function getCachedRxDocumentMonad<RxDocType, OrmMethods>(
                     });
                 }
             });
+            if (docCache.tasks.size <= 1) {
+                requestIdlePromiseNoQueue().then(() => {
+                    docCache.processTasks();
+                });
+            }
         }
         return ret;
     };