+++ src/rx-collection.ts
@@ -218,7 +218,10 @@ export class RxCollectionBase<
         let documentConstructor: any;
         this._docCache = new DocumentCache(
             this.schema.primaryPath,
-            this.$.pipe(filter(cE => !cE.isLocal)),
+            this.database.eventBulks$.pipe(
+                filter(changeEventBulk => changeEventBulk.collectionName === this.name && !changeEventBulk.events[0].isLocal),
+                map(b => b.events)
+            ),
             docData => {
                 if (!documentConstructor) {
                     documentConstructor = getRxDocumentConstructor(this.asRxCollection);
@@ -382,10 +385,11 @@ export class RxCollectionBase<
                 })
             );
         } else {
-            insertRows = [];
+            insertRows = new Array(docsData.length);
+            const schema = this.schema;
             for (let index = 0; index < docsData.length; index++) {
                 const docData = docsData[index];
-                const useDocData = fillObjectDataBeforeInsert(this.schema, docData);
+                const useDocData = fillObjectDataBeforeInsert(schema, docData);
                 insertRows[index] = { document: useDocData };
             }
         }